fix: separated map component with geolocation
